Make sure we don't delete a non existant directory

+++ DomTurnMgr/GameLauncher.cs
@@ -23,7 +23,8 @@ namespace DomTurnMgr
 
     ~GameLauncher()
     {
-      Directory.Delete(tempGameDir, true);
+      if (Directory.Exists(tempGameDir))
+        Directory.Delete(tempGameDir, true);
     }
 
     public Task<string> LaunchAsync()
